Document plugin modules and reorganize core structure
