+++ tsconfig.json
@@ -23,6 +23,6 @@
 		"noUnusedParameters": true,
 		"noPropertyAccessFromIndexSignature": true
 	},
-	"include": ["src/*.ts"],
+	"include": ["src/**/*.ts"],
         "exclude": ["node_modules", "dist", "unit-test"]
 }
