+++ unit-test/store.test.ts
@@ -1,4 +1,4 @@
-import { SnapshotsStore } from '../src/store';
+import { SnapshotsStore } from '../src/core/store';
 import { test, expect } from 'bun:test';
 import { mkdtempSync, readFileSync, readdirSync, rmSync, writeFileSync } from 'node:fs';
 import { tmpdir } from 'node:os';
